:hammer: refactor(game) : 랭킹포인트 int -> Long 으로 변경

+++ backend/src/main/java/com/npc/say_vr/domain/game/domain/Tier.java
@@ -26,5 +26,5 @@ public class Tier {
     private String name;
     private String image;
     private int minPoint;
-    private int maxPoint;
+    private Long maxPoint;
 }
